% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_akmeans.R
\name{akmeans}
\alias{akmeans}
\title{Approximate K-Means Clustering}
\usage{
akmeans(
  x,
  k = 2,
  init = c("Random", "Medoids"),
  maxiter = 20,
  label.init = NULL
)
}
\description{
Approximate K-Means Clustering
}
