% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate diff from best}
\usage{
calc_diff(x)
}
\arguments{
\item{d}{}
}
\value{
Columns with raced differences from the best
}
\description{
Calculate diff from best
}
\examples{
calc_diff(d)
}
