% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cmds.R
\name{cmds}
\alias{cmds}
\title{Classical Multidimensional Scaling}
\usage{
cmds(x, ndim = 2)
}
\description{
Classical Multidimensional Scaling
}
\references{
\insertRef{torgerson_multidimensional_1952}{DAS}
}
\author{
Kisung You
}
