% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cohesion.R
\name{cohesion}
\alias{cohesion}
\title{Cohesion in Finite Metric Space}
\usage{
cohesion(x, label = NULL)
}
\description{
Modularity is included..
}
\references{
\insertRef{chang_mathematical_2016}{DAS}
}
\seealso{
\code{\link{reldist}}
}
