% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cvindex.R
\name{cvindex}
\alias{cvindex}
\title{Cluster Validity Indices based on Pairwise Distance}
\usage{
cvindex(x, label, index = c("Silhouette"))
}
\description{
Cluster Validity Indices based on Pairwise Distance
}
