\name{DAS-package}
\alias{DAS-package}
\alias{DAS}
\docType{package}
\title{
\packageTitle{DAS}
}
\description{
\packageDescription{DAS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DAS}
\packageIndices{DAS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DAS}

Maintainer: \packageMaintainer{DAS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
