% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_emds.R
\name{emds}
\alias{emds}
\title{Euclified Multidimensional Scaling}
\usage{
emds(x, ndim = 2, method = c("closure", "gram"))
}
\description{
strategy 1 : transitive closure of the triangle inequality (labdsv)
strategy 2 : Non-Euclidean or Non-metric Measures Can Be Informative; adding positive numbers to all off-diagonal entries
}
