% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_ksets.R
\name{ksets}
\alias{ksets}
\title{K-Sets Clustering Algorithm}
\usage{
ksets(x, k = 2, init = c("Random", "Medoids"), maxiter = 20)
}
\description{
sort of autotuning is happening..
what about.. just.. recursive partitioning?
}
\references{
\insertRef{chang_mathematical_2016}{DAS}
}
