% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_mds.R
\name{mds}
\alias{mds}
\title{Classical Multidimensional Scaling}
\usage{
mds(x, ndim = 2)
}
\description{
Classical Multidimensional Scaling
}
\references{
\insertRef{kruskal_multidimensional_1964}{DAS}
}
\author{
Kisung You
}
