% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_nef.R
\name{nef}
\alias{nef}
\title{Negative Eigenfraction}
\usage{
nef(x)
}
\description{
NEF == 0; data is euclidean
NEF >> 0; far
}
\details{
Gram matrix -1/2 JD^2J be positive semidefinite == EUclidean
}
\examples{
## use simple example of iris dataset 
data(iris) 
dat  = iris[,1:4]
dmat = as.matrix(stats::dist(dat))

## run the algorithm
nef(dmat)

}
\references{
\insertRef{pekalska_non-euclidean_2006}{DAS}
}
\seealso{
\code{\link{nem}}
}
\author{
Kisung You
}
