% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_nem.R
\name{nem}
\alias{nem}
\title{Negative Eigenvalue Magnitude}
\usage{
nem(x)
}
\description{
NEM == 0; data is euclidean
NEM >> 0; far
}
\details{
Gram matrix -1/2 JD^2J be positive semidefinite == EUclidean
}
\examples{
## use simple example of iris dataset 
data(iris) 
dat  = iris[,1:4]
dmat = as.matrix(stats::dist(dat))

## run the algorithm
nem(dmat)

}
\references{
\insertRef{pekalska_non-euclidean_2006}{DAS}
}
\seealso{
\code{\link{nef}}
}
\author{
Kisung You
}
