% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opaque.R
\name{opaque}
\alias{opaque}
\title{Make a color opaque}
\usage{
opaque(color, opacity, max = 255)
}
\arguments{
\item{color}{Name or hex code of color}

\item{opacity}{Opacity (default is out of 255)}

\item{max}{Maximum value (typically 1 or 255)}
}
\description{
Make a color opaque
}
\examples{
opaque('red', 50)
}
\keyword{color}
