% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek at an object of any type}
\usage{
peek(obj, nrowToShow = 3, ncolToShow = 4)
}
\arguments{
\item{obj}{The object you wish to explore}

\item{nrowToShow}{Number of rows to show, defaults to 3}

\item{ncolToShow}{Number of cols to show, defaults to 4}
}
\description{
This is an ugly function that I have found useful for looking at
objects of various types
}
\examples{
peek(matrix(1:12, nrow=3), nrow=2, ncol=2)
}
