% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_reldist.R
\name{reldist}
\alias{reldist}
\title{Relative Distance in Finite Metric Space}
\usage{
reldist(x, label = NULL)
}
\description{
Relative Distance in Finite Metric Space
}
\references{
\insertRef{chang_mathematical_2016}{DAS}
}
