% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_testk.mdf.R
\name{testk.mdf}
\alias{testk.mdf}
\title{MDF test}
\usage{
testk.mdf(x, label, mc.iter = 999)
}
\arguments{
\item{x}{an \eqn{(N\times N)} matrix or \code{dist} object.}

\item{label}{length-\eqn{N} vector of class/grouping label.}

\item{mc.iter}{the number of permutations to be taken.}
}
\value{
a (list) object of \code{S3} class \code{htest} containing: \describe{
\item{statistic}{a test statistic.}
\item{p.value}{\eqn{p}-value under \eqn{H_0}.}
\item{alternative}{alternative hypothesis.}
\item{method}{name of the test.}
\item{data.name}{name(s) of provided sample data.}
\item{permvec}{statistics from each permutation.}
}
}
\description{
Given the distance matrix of two or more samples, it tests
\deqn{H_0 : F_{g_1} = \cdots = F_{g_K} F_Y\quad vs\quad H_1 : \textrm{ not } H_0}
using the procedure by Minas et al. (2011) in a nonparametric way 
that depends on \code{mc.iter} number of permutations.
}
\examples{
## small test for CRAN submission
dat1 <- matrix(rnorm(60, mean= 1), ncol=2) # group 1 : 30 obs of mean  1
dat2 <- matrix(rnorm(50, mean=-1), ncol=2) # group 2 : 25 obs of mean -1

dmat <- as.matrix(dist(rbind(dat1, dat2)))  # Euclidean distance matrix
lab  <- c(rep(1,30), rep(2,25))             # corresponding label

testk.mdf(dmat, lab)                        # run the code !

\dontrun{
## WARNING: computationally heavy. 
#  Let's compute empirical Type 1 error at alpha=0.05
niter = 496  
pvals = rep(0,niter)
for (i in 1:niter){
  dat = matrix(rnorm(200),ncol=2)
  lab = c(rep(1,50), rep(2,50))
  pvals[i] = testk.mdf(as.matrix(dist(dat)), lab)$p.value
  print(paste("iteration ",i," complete..",sep=""))
}
print(paste("* Empirical Type 1 Error : ",sum(pvals<=0.05)/niter,sep=""))

#  Visualize the above at multiple significance levels
alphas = seq(from=0.001, to=0.999, length.out=100)
errors = rep(0,100)
for (i in 1:100){
   errors[i] = sum(pvals<=alphas[i])/niter
}
plot(alphas, errors, "b", main="Empirical Type 1 Errors", 
     xlab="alpha", ylab="error", lwd=2)
abline(v=0.05, lwd=2, col="red")
}

}
\references{
\insertRef{minas_distance-based_2011}{DAS}
}
