% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_excel}
\alias{write_excel}
\title{A Function to Write Excel Spreadsheets}
\usage{
write_excel(x, xlfile)
}
\arguments{
\item{x}{data.frame or list of data.frames to write to file}

\item{xlfile}{path to Excel file you want to write}
}
\description{
Write a data.frame or list of them to an Excel spreadsheet
Auto-widen columns for easy browsing.
}
\examples{
write_excel()
}
