% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek at an object of any type}
\usage{
peek(obj, n_row_to_show = 3, n_col_to_show = 4)
}
\arguments{
\item{obj}{The object you wish to explore}

\item{n_row_to_show}{Number of rows to show, defaults to 3}

\item{n_col_to_show}{Number of cols to show, defaults to 4}
}
\description{
This is an ugly function that I have found useful for looking at
objects of various types
}
\examples{
peek(matrix(1:12, nrow=3), n_row=2, n_col=2)
}
